+++ packages/@cldcvr/flow-form-builder/src/components/f-form-array/f-form-array.ts
@@ -226,7 +226,7 @@ export class FFormArray extends FRoot {
 	 * dispatch showWhen event so that root will publish new form values
 	 */
 	dispatchShowWhenEvent() {
-		const showWhen = new CustomEvent("showWhen", {
+		const showWhen = new CustomEvent("show-when", {
 			detail: true,
 			bubbles: true,
 			composed: true