+++ packages/@cldcvr/flow-form-builder/src/components/f-form-builder/fields/text.ts
@@ -15,16 +15,17 @@ export default function (
 			${ref(fieldRef)}
 			.placeholder=${field.placeholder}
 			.type=${field.type}
-			icon-left=${ifDefined(field?.iconLeft)}
-			icon-right=${ifDefined(field?.iconRight)}
-			prefix=${ifDefined(field?.prefix)}
-			suffix=${ifDefined(field?.suffix)}
-			state=${ifDefined(field?.state)}
-			max-length=${ifDefined(field?.maxLength)}
-			?loading=${field?.loading ?? false}
-			?disabled=${field?.disabled ?? false}
-			?clear=${field?.clear ?? true}
-			?read-only=${field?.readonly ?? false}
+			icon-left=${ifDefined(field.iconLeft)}
+			icon-right=${ifDefined(field.iconRight)}
+			prefix=${ifDefined(field.prefix)}
+			suffix=${ifDefined(field.suffix)}
+			state=${ifDefined(field.state)}
+			.suffixWhen=${field.suffixWhen}
+			max-length=${ifDefined(field.maxLength)}
+			?loading=${field.loading ?? false}
+			?disabled=${field.disabled ?? false}
+			?clear=${field.clear ?? true}
+			?read-only=${field.readonly ?? false}
 			@click=${ifDefined(field.onClick)}
 			@focus=${ifDefined(field.onFocus)}
 			@input=${ifDefined(field.onInput)}